+++ app/controllers/api/sections_controller.rb
@@ -1,6 +1,17 @@
-# A controller responsible for CRUD operations for the Section model
 module Api
+  # Api controller responsible for CRUD operations for the Section model
   class SectionsController < MainApiController
+    def create
+      section = current_course.sections.new(section_params)
+      if section.save
+        render 'shared/http_status',
+               locals: { code: '201', message: HttpStatusHelper::ERROR_CODE['message']['201'] }, status: :created
+      else
+        render 'shared/http_status', locals: { code: '422', message:
+          HttpStatusHelper::ERROR_CODE['message']['422'] }, status: :unprocessable_entity
+      end
+    end
+
     def destroy
       @section = record
 
@@ -13,6 +24,10 @@ module Api
         render 'shared/http_status',
                locals: { code: '200', message: t('sections.destroy.success'), status: :ok }
       end
+    end
+
+    def section_params
+      params.require(:section).permit(:name)
     end
   end
 end